pyproject.toml
[project]
name = "end-to-end-rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt==0.9",
    "faiss-cpu>=1.13.1",
    "fastapi>=0.127.0",
    "google-genai>=1.56.0",
    "ipykernel==6.30.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "langchain-google-vertexai>=3.2.0",
    "langchain-groq>=1.1.1",
    "langchain-text-splitters>=1.1.0",
    "langsmith>=0.5.0",
    "openevals>=0.1.3",
    "pandas>=2.3.3",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart==0.0.20",
    "rapidocr-onnxruntime>=1.4.4",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]
